create a lark-test crate
So we don't have to recreate the query system
