+++ components/lark-hir/src/lib.rs
@@ -24,7 +24,6 @@ use std::sync::Arc;
 mod fn_body;
 mod query_definitions;
 mod scope;
-mod test;
 mod type_conversion;
 
 salsa::query_group! {
